+++ tests/test_integration.py
@@ -49,6 +49,47 @@ def test_train_and_run(
     assert model_file.exists()
     assert best_model.exists()
 
+    # Resume training from previous checkpoint.
+    with tiny_config.open("r") as f:
+        config_data = yaml.safe_load(f)
+    original_max_epochs = config_data.get("max_epochs")
+
+    try:
+        # max_epochs needs to be increased to train further.
+        if original_max_epochs is not None:
+            config_data["max_epochs"] = original_max_epochs + 10
+        else:
+            config_data["max_epochs"] = 10
+        with tiny_config.open("w") as f:
+            yaml.dump(config_data, f)
+
+        train_args = [
+            "train",
+            str(mgf_small),
+            "--config",
+            tiny_config,
+            "--output_dir",
+            str(tmp_path),
+            "--output_root",
+            "train_resuming",
+            "--model",
+            str(model_file),
+            "--load_all_states",
+        ]
+
+        result = run(train_args)
+        best_model = tmp_path / "train_resuming.best.ckpt"
+        assert result.exit_code == 0
+        assert best_model.exists()
+    finally:
+        if original_max_epochs is None:
+            config_data.pop("max_epochs", None)
+        else:
+            config_data["max_epochs"] = original_max_epochs
+
+        with tiny_config.open("w") as f:
+            yaml.dump(config_data, f)
+
     # Run Casanovo in de novo prediction mode.
     output_rootname = "test"
     output_filename = (tmp_path / output_rootname).with_suffix(".mztab")